FaceGenerator methods now take the vertices and edge vectors as parameters
generateTrianglesList,generateTriangleEdgePairs, and calculateMaxWidth now take
the face's vertices and edge index vectors as parameters rather than storing
them as members of the FaceGenerator class. FaceGenerator's constructor no longer
takes any parameters as a result of the change. All subclasses of FaceGenerator
and the Polyhedron class have been updated to reflect this change.

This change was made in order to remove the FaceVertices object from FaceGenerator.
Removing the FaceVertices member allows FaceGenerators to exist independently of a
polyhedron, which was impossible before. The FaceVertices class was used to get
the values of the FaceGenerator's vertices from the polyhedron it was a part of.
This task is now accomplished by passing the relevant vertex values as parameters
to FaceGenerator methods.